refactor: move core atoms in a core directory
