+++ src/atoms/core/atom.types.ts
@@ -1,0 +1,8 @@
+export type PaneId = 'LEFT' | 'RIGHT';
+
+export type FileId = string;
+
+export interface PaneFileId {
+  paneId: PaneId;
+  fileId: FileId;
+}
