+++ src/types/FileEntry.ts
@@ -4,13 +4,13 @@ interface FileEntryBase {
   isDotfile: boolean;
 }
 
-interface FileFileEntry extends FileEntryBase {
+export interface FileFileEntry extends FileEntryBase {
   isFile: true;
   isFolder: false;
   fileExtension: string;
 }
 
-interface FolderFileEntry extends FileEntryBase {
+export interface FolderFileEntry extends FileEntryBase {
   isFile: false;
   isFolder: true;
   children: FileEntry[];
